test(dsl): enhance Flow DSL type system tests and utility types
Refactor and expand type system tests for Flow DSL, adding comprehensive type checking for:
- Step dependencies and payload type constraints
- Multi-level dependency type inference
- StepOutput utility type with complex nested type support
- Compile-time type safety checks for step dependencies

Remove redundant test file and consolidate type system tests in a single file with improved
test coverage
