--- pkgs/dsl/src/failing02.test-d.ts
@@ -1,10 +1,0 @@
-import { test, expectTypeOf } from 'vitest';
-
-// This test should fail type checking
-test('should explicitly fail type checking', () => {
-  // @ts-expect-error
-  const x: string = 42; // This should cause a type error
-
-  // This should fail type checking
-  expectTypeOf<number>().toEqualTypeOf<string>();
-});
